Got chunks generating at the right positions but scale is broken. All chunks generating at the same size. May rewrite PlaneMeshGenerator to me more intuitive.

+++ Assets/_Scripts/QuadChunk.cs
@@ -8,7 +8,7 @@ public class QuadChunk {
     float sideLength;
     int lodLevel;
 
-    public QuadChunk(Vector3 botLeftPoint, float lodConstant, float sideLength) {
+    public QuadChunk(Vector3 botLeftPoint, int lodConstant, float sideLength) {
         this.botLeftPoint = botLeftPoint;
         this.sideLength = sideLength;
         //this.lodConstant = lodConstant;
